moving common utility functions out of __init__
